Struct str_index::StrIndex

pub struct StrIndex { /* fields omitted */ }

An index into a string.

The index is stored as a 32 bit integer, assuming we only deal with text shorter than 4 GiB.

Methods

impl StrIndex

pub fn from_char_len(c: char) -> Self

Index equal to the string length of this char.

Examples

assert_eq!(
    StrIndex::from_char_len('😂'),
    StrIndex::from(4),
);

pub fn from_str_len(s: &str) -> Self

Index equal to the length of this string.

Examples

assert_eq!(
    StrIndex::from_str_len("メカジキ"),
    StrIndex::from(12),
);

pub fn to_usize(self) -> usize

This index as a raw usize.

pub fn checked_add(self, rhs: Self) -> Option<Self>

Checked integer addition.

pub fn checked_sub(self, rhs: Self) -> Option<Self>

Checked integer subtraction.

pub fn range_for(self, len: StrIndex) -> StrRange

A range starting at this index.

Example

let point = StrIndex::from(5);
assert_eq!(
    point.range_for(10.into()),
    StrRange::from(5.into()..15.into()),
);

pub fn range_to(self, end: StrIndex) -> StrRange

A range from this index to another.

Panics

Panics if end < self.

Example

let start = StrIndex::from(0);
let end = StrIndex::from(10);
assert_eq!(
    start.range_to(end),
    StrRange::from(start..end),
);

pub fn as_unit_range(self) -> StrRange

The empty range at this index.

Example

let point = StrIndex::from(10);
assert_eq!(
    point.as_unit_range(),
    point.range_to(point),
);